pyproject.toml
[project]
name = "backtrader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "matplotlib>=3.10.5",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.1",
    "uvicorn>=0.35.0",
    "alpaca-trade-api>=3.2.0",
    "yfinance>=0.2.58",
]

[dependency-groups]
dev = [
    "backtrader",
]

[tool.uv.sources]
backtrader = { workspace = true }
